Cert subscriber count to Dev (#82) (#85)
* Cert subscriber count (#81)

* fix: Delete __tests__/setup/globalConfigMongo.json

- setup json file deleted to stop tracking of the file

* fix(test): Deleted some leftover commented-out code

* fix: updated to new error messages and fixed auth route types

* fix(api): deleted some leftover duplicate code

* fix: (DIRECT COMMIT) removed consolelogs

- Removed console log from adminOnly.test.js
- Removed console log from contentCreator.test.js
- Removed console log from courseRoutes.test.js
- Removed console log from userRoutes.js

* fix:(DIRECT COMMIT) changed delete route

- Changed delete route endpoint for users in userRoutes.js
- Changed tests to follow new endpoint in userRoutes.test.js

* fix: (DIRECT COMMIT) pluralized user endpoint

- Changed user endpoint to pluralized version to follow guidelines
- Changed endpoints in tests to compensate.

* Fix: removed password from token

and replaced it with _id

* feat(api): Added subscriber count functionality

Changes made:
- Added two error codes:
    -  Attempting to unsubscribe from a course you are not subscribed to
    - Attempting to subscribe to a course you are already subscribed to
- Updated subscribe routes to send the updated user instead of the old one
- Added an update to the coursemodels with an increment or decrement of one to the subscribe count for subscribe and unsubscribe respectively

* doc: added TODO for sub count

* fix(course tests) refactored tests

- removed one test that only tested response, and not actual database object
- added tests for unsubscribing, and subscribing that tests both numOfSubscriptions for courses, and subscriptions for users
- added test for unsubscribing to a test without any subscribers
- added numOfSubscriptions: 0 as fakeCourse object did not specify this amount

---------






* feat(courseRoutes test) added another test

- added test to check error handling if user attempts to subscribe to a course, that user is already subscribed to

* fix(courses) fixed get courses

- does not return error, when no courses are found, as it now just returns empty course list
- updated test to resemble above

* fix(tests) merged two test files to one

- merged courseroutes.spec.js and courseroutes.test.js
- updated the fakeCourse and fakeCourses model to slighly reflect the actual model more
- updated the tests to reflect the updated model

* fixed connectDB import

* fix(courseRoutes): dangling catch

- removed duplicated misplaced catch

---------
