+++ routes/courseRoutes.js
@@ -2,6 +2,8 @@ const router = require('express').Router();
 const errorCodes = require('../helpers/errorCodes');
 const adminOnly = require("../middlewares/adminOnly");
 
+// TODO: Update subscriber count to check actual value in DB
+
 // Models
 const { CourseModel } = require('../models/Courses');
 const { SectionModel } = require('../models/Sections');
@@ -9,7 +11,7 @@ const { ComponentModel } = require('../models/Components');
 const { ExerciseModel } = require('../models/Exercises');
 const { UserModel } = require('../models/Users');
 const {
-	ContentCreatorApplication,
+  ContentCreatorApplication,
 } = require("../models/ContentCreatorApplication");
 const requireLogin = require("../middlewares/requireLogin");
 const { IdentityStore } = require("aws-sdk");
@@ -20,106 +22,109 @@ const { IdentityStore } = require("aws-sdk");
 router.get('/creator/:id', requireLogin, async (req, res) => {
   const id = req.params.id; // Get user id from request
   const courses = await CourseModel.find({ creator: id }); // Find courses for a specific user
-	
+
+
+
   res.send(courses); // Send response
 });
 
 //Get all courses
 router.get('/', async (req, res) => {
 
-	try {
-		// find all courses in the database
-		const courses = await CourseModel.find();
-		res.send(courses);
-	} catch (error) {
-		// If the server could not be reached, return an error message
-		return res.status(500).json({ 'error': errorCodes['E0003'] });
-	}
+  try {
+    // find all courses in the database
+    const courses = await CourseModel.find();
+    res.send(courses);
+  } catch (error) {
+    // If the server could not be reached, return an error message
+    return res.status(500).json({ 'error': errorCodes['E0003'] });
+  }
 });
-
 
 // Get specific course
 router.get('/:id', async (req, res) => {
-	try {
-		const { id } = req.params;
+  try {
+    const { id } = req.params;
 
-		// find a course based on it's id
-		const course = await CourseModel.findById(id);
+    // find a course based on it's id
+    const course = await CourseModel.findById(id);
 
-		// check if courses exist
-		if (!course) {
-			// Handle "course not found" error response here
-			return res.status(404).json({ 'error': errorCodes['E0006'] });
-		}
-		res.send(course);
-	} catch (error) {
-		return res.status(500).json({ 'error': errorCodes['E0003'] });
-	}
+    // check if courses exist
+    if (!course) {
+      // Handle "course not found" error response here
+      return res.status(404).json({ 'error': errorCodes['E0006'] });
+    }
+    res.send(course);
+  } catch (error) {
+    return res.status(500).json({ 'error': errorCodes['E0003'] });
+  }
 });
 
 // Get all sections from course
 router.get('/:id/sections', async (req, res) => {
-	try {
-		const { id } = req.params;
 
-		// find a course based on it's id
-		const course = await CourseModel.findById(id);
+  try {
+    const { id } = req.params;
 
-		// check if courses exist
-		if (!course) {
-			// Handle "course not found" error response here
-			return res.status(404).json({ 'error': errorCodes['E0006'] });
-		}
 
-		const sectionsInCourse = course.sections;
+    // find a course based on it's id
+    const course = await CourseModel.findById(id);
 
-		// check if course contains sections
-		if (sectionsInCourse.length === 0) {
-			// Handle "course does not contain sections" error response here
-			return res.status(404).json({ 'error': errorCodes['E0009'] });
-		}
+    // check if courses exist
+    if (!course) {
+      // Handle "course not found" error response here
+      return res.status(404).json({ 'error': errorCodes['E0006'] });
+    }
 
-		// check if the sections exists
-		const sectionsList = await SectionModel.find({ '_id': { $in: sectionsInCourse } });
-		if (sectionsList.length === 0) {
-			// Handle "course does not contain sections" error response here
-			return res.status(404).json({ 'error': errorCodes['E0007'] });
-		}
+    const sectionsInCourse = course.sections;
 
-		res.send(sectionsList);
+    // check if course contains sections
+    if (sectionsInCourse.length === 0) {
+      // Handle "course does not contain sections" error response here
+      return res.status(404).json({ 'error': errorCodes['E0009'] });
+    }
 
-	} catch (error) {
-		return res.status(500).json({ 'error': errorCodes['E0003'] });
-	}
+    // check if the sections exists
+    const sectionsList = await SectionModel.find({ '_id': { $in: sectionsInCourse } });
+    if (sectionsList.length === 0) {
+      // Handle "course does not contain sections" error response here
+      return res.status(404).json({ 'error': errorCodes['E0007'] });
+    }
+
+    res.send(sectionsList);
+
+  } catch (error) {
+    return res.status(500).json({ 'error': errorCodes['E0003'] });
+  }
 
 });
 
 // Get a specififc section 
 router.get('/:courseId/sections/:sectionId', async (req, res) => {
 
-	try {
-		const { courseId, sectionId } = req.params;
+  try {
+    const { courseId, sectionId } = req.params;
 
-		const course = await CourseModel.findById(courseId);
-		// check if courses exist
-		if (!course) {
-			// Handle "course not found" error response here
-			return res.status(404).json({ 'error': errorCodes['E0006'] });
-		}
-		// find a specific section within the given course by both IDs
-		const section = await SectionModel.findOne({ parentCourse: courseId, _id: sectionId });
+    const course = await CourseModel.findById(courseId);
+    // check if courses exist
+    if (!course) {
+      // Handle "course not found" error response here
+      return res.status(404).json({ 'error': errorCodes['E0006'] });
+    }
+    // find a specific section within the given course by both IDs
+    const section = await SectionModel.findOne({ parentCourse: courseId, _id: sectionId });
 
-		// check if section exist
-		if (!section) {
-			// Handle "section not found" error response here
-			return res.status(404).json({ 'error': errorCodes['E0008'] });
-		}
+    // check if section exist
+    if (!section) {
+      // Handle "section not found" error response here
+      return res.status(404).json({ 'error': errorCodes['E0008'] });
+    }
 
-		res.send(section);
+    res.send(section);
 
-	} catch (error) {
-		return res.status(500).json({ 'error': errorCodes['E0003'] });
-	}
+  } catch (error) {
+    return res.status(500).json({ 'error': errorCodes['E0003'] });
+  }
 });
 
 /*** SUBSCRIPTION ROUTES ***/
@@ -127,80 +132,107 @@ router.get('/:courseId/sections/:sectionId', async (req, res) => {
 // Subscribe to course 
 router.post('/:id/subscribe', async (req, res) => {
 
-	try {
-		const { id } = req.params;
-		const { user_id } = req.body;
+  try {
+    const { id } = req.params;
+    const { user_id } = req.body;
 
-		const user = await UserModel.findById(user_id);
+    const user = await UserModel.findById(user_id);
 
-		//checks if user exist
-		if (!user) {
-			// Handle "user not found" error response here
-			return res.status(404).json({ 'error': errorCodes['E0004'] });
-		}
+    //checks if user exist
+    if (!user) {
+      // Handle "user not found" error response here
+      return res.status(404).json({ 'error': errorCodes['E0004'] });
+    }
 
-		const course = await CourseModel.findById(id);
-		// check if courses exist
-		if (!course) {
-			// Handle "course not found" error response here
-			return res.status(404).json({ 'error': errorCodes['E0006'] });
-		}
+    const course = await CourseModel.findById(id);
+    // check if courses exist
+    if (!course) {
+      // Handle "course not found" error response here
+      return res.status(404).json({ 'error': errorCodes['E0006'] });
+    }
 
-		// find user based on id, and add the course's id to the user's subscriptions field
-		(await UserModel.findOneAndUpdate(
-			{ _id: user_id },
-			{ $push: { subscriptions: id } }))
-			.save;
+    if (user.subscriptions.includes(id)) {
+      return res.status(400).json({ 'error': errorCodes['E0605'] }); //TODO: change error code
+    }
 
-		res.send(user);
+    course.numOfSubscriptions++;
+    user.subscriptions.push(id);
 
-	} catch (error) {
-		return res.status(500).json({ 'error': errorCodes['E0003'] });
-	}
+
+    // find user based on id, and add the course's id to the user's subscriptions field
+    await UserModel.findOneAndUpdate(
+      { _id: user_id },
+      { subscriptions: user.subscriptions });
+
+    await CourseModel.findOneAndUpdate(
+      { _id: course._id },
+      { numOfSubscriptions: course.numOfSubscriptions }
+    );
+
+
+    res.status(200).send(user);
+
+  } catch (error) {
+    console.log(error);
+    return res.status(500).json({ 'error': errorCodes['E0003'] });
+  }
 
 });
 
 // Unsubscribe to course
 router.post('/:id/unsubscribe', async (req, res) => {
 
-	try {
-		const { id } = req.params;
-		const { user_id } = req.body;
+  try {
+    const { id } = req.params;
+    const { user_id } = req.body;
 
-		const user = await UserModel.findById(user_id);
-		//checks if user exist
-		if (!user) {
-			// Handle "user not found" error response here
-			return res.status(404).json({ 'error': errorCodes['E0004'] });
-		}
+    const user = await UserModel.findById(user_id);
+    //checks if user exist
+    if (!user) {
+      // Handle "user not found" error response here
+      return res.status(404).json({ 'error': errorCodes['E0004'] });
+    }
 
-		const course = await CourseModel.findById(id);
-		// check if courses exist
-		if (!course) {
-			// Handle "course not found" error response here
-			return res.status(404).json({ 'error': errorCodes['E0006'] });
-		}
+    const course = await CourseModel.findById(id);
+    // check if courses exist
+    if (!course) {
+      // Handle "course not found" error response here
+      return res.status(404).json({ 'error': errorCodes['E0006'] });
+    }
 
-		// find user based on id, and remove the course's id from the user's subscriptions field
-		(await UserModel.findOneAndUpdate(
-			{ _id: user_id },
-			{ $pull: { subscriptions: id } }))
-			.save;
+    if (!user.subscriptions.includes(id)) {
+      return res.status(400).json({ 'error': errorCodes['E0606'] }); //TODO: change error code
+    }
 
-		res.send(user)
+    course.numOfSubscriptions--;
+    // Remove course id from user's subscriptions
+    user.subscriptions.indexOf(id) > -1 && user.subscriptions.splice(user.subscriptions.indexOf(id), 1);
 
-	} catch (error) {
-		return res.status(500).json({ 'error': errorCodes['E0003'] });
-	}
+    // find user based on id, and remove the course's id from the user's subscriptions field
+    await UserModel.findOneAndUpdate(
+      { _id: user_id },
+      { subscriptions: user.subscriptions })
+
+    await CourseModel.findOneAndUpdate(
+      { _id: course._id },
+      { numOfSubscriptions: course.numOfSubscriptions }
+    );
+
+    res.send(user)
+
+  } catch (error) {
+    console.log(error);
+    return res.status(500).json({ 'error': errorCodes['E0003'] });
+  }
 
 });
 
 // Get all exercises for section
 router.get('/:section_id/exercises', async (req, res) => {
-	const { section_id } = req.params;
-	const section = await SectionModel.findById(section_id);
-	const list = await ExerciseModel.find({ _id: section.exercises });
-	res.send(list);
+  const { section_id } = req.params;
+  const section = await SectionModel.findById(section_id);
+  const list = await ExerciseModel.find({ _id: section.exercises });
+  res.send(list);
 });
 
 /*** CREATE COURSE ROUTES ***/
@@ -208,7 +240,7 @@ router.get('/:section_id/exercises', async (req, res) => {
 //Create course route
 router.put("/", async (req, res) => {
   const { title, category, difficulty, description, estimatedHours } = req.body;
-	console.dir(req.body);
+  console.dir(req.body);
   const course = new CourseModel({
     title: title,
     category: category,
@@ -221,12 +253,12 @@ router.put("/", async (req, res) => {
     dateUpdated: Date.now(),
     sections: [],
     estimatedHours: estimatedHours,
-	rating: 0,
+    rating: 0,
   });
 
   try {
     await course.save();
-	res.status(201).send(course);
+    res.status(201).send(course);
   } catch (err) {
     res.status(400).send(err);
   }
@@ -240,18 +272,18 @@ router.patch("/:id", /*requireLogin,*/ async (req, res) => {
   const dbCourse = await CourseModel.findByIdAndUpdate(
     id,
     {
-	  title: course.title,
-	  description: course.description,
-	  category: course.category,
-	  difficulty: course.difficulty,
-	  estimatedHours: course.estimatedHours,
+      title: course.title,
+      description: course.description,
+      category: course.category,
+      difficulty: course.difficulty,
+      estimatedHours: course.estimatedHours,
       published: course.published,
-	  dateUpdated: Date.now()
+      dateUpdated: Date.now()
     },
     function (err, docs) {
       if (err) {
         res.status(400).send(err);
-      } 
+      }
     }
   );
   res.status(200).send(dbCourse);
@@ -265,7 +297,7 @@ router.patch("/:id", /*requireLogin,*/ async (req, res) => {
  * 
  * @param {string} id - course id
  * @returns {string} - Just sends a message to confirm that the deletion is complete
- */ 
+ */
 router.delete("/:id"/*, requireLogin*/, async (req, res) => {
   const { id } = req.params;
 
@@ -285,7 +317,7 @@ router.delete("/:id"/*, requireLogin*/, async (req, res) => {
 
     // Get the lecture array from the section object
     const lectureIds = section.lectures;
-	const exerciseIds = section.exercises;
+    const exerciseIds = section.exercises;
 
     // Loop through all lectures in section
     lectureIds.map(async (lecture_id) => {
@@ -295,13 +327,13 @@ router.delete("/:id"/*, requireLogin*/, async (req, res) => {
 
     });
 
-	// Loop through all exercises in section
-	exerciseIds.map(async (exercise_id) => {
+    // Loop through all exercises in section
+    exerciseIds.map(async (exercise_id) => {
 
-		// Delete the exercise
-		await ExerciseModel.findByIdAndDelete(exercise_id);
+      // Delete the exercise
+      await ExerciseModel.findByIdAndDelete(exercise_id);
 
-	});
+    });
 
     // Delete the section
     await SectionModel.findByIdAndDelete(section_id);
